fix: preserve crop affine precision

+++ composeApp/src/commonMain/kotlin/presentation/screens/gallery/editor/CropRenderPlanner.kt
@@ -12,21 +12,21 @@ class CropRenderPlanner(
         val geometry = calculator.geometry(source, output, request.transform)
         val turns = normalizeTurns(request.transform.quarterTurns)
         val uniformScale = if (turns % 2 == 0) {
-            geometry.imageWidth / source.width
+            geometry.imageWidth.toDouble() / source.width
         } else {
-            geometry.imageHeight / source.width
+            geometry.imageHeight.toDouble() / source.width
         }
-        val flipX = geometry.scaleXSign
-        val flipY = geometry.scaleYSign
+        val flipX = geometry.scaleXSign.toDouble()
+        val flipY = geometry.scaleYSign.toDouble()
         val rotation = rotation(turns)
         val scaleX = rotation.scaleX * uniformScale * flipX
         val skewX = rotation.skewX * uniformScale * flipY
         val skewY = rotation.skewY * uniformScale * flipX
         val scaleY = rotation.scaleY * uniformScale * flipY
-        val outputCenterX = output.width / 2f + geometry.translationX
-        val outputCenterY = output.height / 2f + geometry.translationY
-        val sourceCenterX = source.width / 2f
-        val sourceCenterY = source.height / 2f
+        val outputCenterX = output.width / 2.0 + geometry.translationX.toDouble()
+        val outputCenterY = output.height / 2.0 + geometry.translationY.toDouble()
+        val sourceCenterX = source.width / 2.0
+        val sourceCenterY = source.height / 2.0
         val sourceToOutput = AffineTransform(
             scaleX = scaleX,
             skewX = skewX,
@@ -48,12 +48,12 @@ class CropRenderPlanner(
         source: ImageSize,
         output: ImageSize,
     ): PixelRect {
-        val scaleX = transform.scaleX.toDouble()
-        val skewX = transform.skewX.toDouble()
-        val translateX = transform.translateX.toDouble()
-        val skewY = transform.skewY.toDouble()
-        val scaleY = transform.scaleY.toDouble()
-        val translateY = transform.translateY.toDouble()
+        val scaleX = transform.scaleX
+        val skewX = transform.skewX
+        val translateX = transform.translateX
+        val skewY = transform.skewY
+        val scaleY = transform.scaleY
+        val translateY = transform.translateY
         val determinant = scaleX * scaleY - skewX * skewY
         val outputCorners = listOf(
             0.0 to 0.0,
@@ -87,18 +87,18 @@ class CropRenderPlanner(
         }
 
     private fun rotation(turns: Int): LinearTransform = when (turns) {
-        0 -> LinearTransform(1f, 0f, 0f, 1f)
-        1 -> LinearTransform(0f, -1f, 1f, 0f)
-        2 -> LinearTransform(-1f, 0f, 0f, -1f)
-        else -> LinearTransform(0f, 1f, -1f, 0f)
+        0 -> LinearTransform(1.0, 0.0, 0.0, 1.0)
+        1 -> LinearTransform(0.0, -1.0, 1.0, 0.0)
+        2 -> LinearTransform(-1.0, 0.0, 0.0, -1.0)
+        else -> LinearTransform(0.0, 1.0, -1.0, 0.0)
     }
 
     private fun normalizeTurns(turns: Int): Int = ((turns % 4) + 4) % 4
 
     private data class LinearTransform(
-        val scaleX: Float,
-        val skewX: Float,
-        val skewY: Float,
-        val scaleY: Float,
+        val scaleX: Double,
+        val skewX: Double,
+        val skewY: Double,
+        val scaleY: Double,
     )
 }
